requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
python-decouple>=3.8
pathlib>=1.0.1
typing-extensions>=4.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
